// src/platform/plugins/shared/chart_expressions/expression_xy/public/types.d.ts
import type { IconType } from '@elastic/eui';
import type { DataPublicPluginSetup } from '@kbn/data-plugin/public';
import type { FieldFormatsSetup } from '@kbn/field-formats-plugin/public';
import type { ChartsPluginSetup } from '@kbn/charts-plugin/public';
import type { IFieldFormat, SerializedFieldFormat } from '@kbn/field-formats-plugin/common';
import type { CellValueContext, RangeSelectContext, ValueClickContext, MultiValueClickContext } from '@kbn/embeddable-plugin/public';
import type { ExpressionsServiceStart, ExpressionsSetup } from '@kbn/expressions-plugin/public';
import type { CoreStart } from '@kbn/core/public';
export interface SetupDeps {
    expressions: ExpressionsSetup;
    data: DataPublicPluginSetup;
    fieldFormats: FieldFormatsSetup;
    charts: ChartsPluginSetup;
}
export interface StartDeps {
    expression: ExpressionsServiceStart;
}
export type StartServices = Pick<CoreStart, 'analytics' | 'i18n' | 'theme' | 'userProfile'>;
export type ExpressionXyPluginSetup = void;
export type ExpressionXyPluginStart = void;
export interface FilterEvent {
    name: 'filter';
    data: ValueClickContext['data'];
}
export interface MultiFilterEvent {
    name: 'multiFilter';
    data: MultiValueClickContext['data'];
}
export interface BrushEvent {
    name: 'brush';
    data: RangeSelectContext['data'];
}
export type FormatFactory = (mapping?: SerializedFieldFormat) => IFieldFormat;
export interface OperationDescriptor extends Operation {
    hasTimeShift: boolean;
}
export type SortingHint = 'version';
export type FieldOnlyDataType = 'document' | 'ip' | 'histogram' | 'geo_point' | 'geo_shape';
export type DataType = 'string' | 'number' | 'date' | 'boolean' | FieldOnlyDataType;
export interface OperationMetadata {
    dataType: DataType;
    isBucketed: boolean;
    /**
     * ordinal: Each name is a unique value, but the names are in sorted order, like "Top values"
     * interval: Histogram data, like date or number histograms
     * ratio: Most number data is rendered as a ratio that includes 0
     */
    scale?: 'ordinal' | 'interval' | 'ratio';
    isStaticValue?: boolean;
}
export interface Operation extends OperationMetadata {
    label: string;
    sortingHint?: SortingHint;
}
/**
 * A visualization type advertised to the user in the chart switcher
 */
export interface VisualizationType {
    /**
     * Unique id of the visualization type within the visualization defining it
     */
    id: string;
    /**
     * Icon used in the chart switcher
     */
    icon: IconType;
    /**
     * Visible label used in the chart switcher and above the workspace panel in collapsed state
     */
    label: string;
    /**
     * Optional label used in visualization type search if chart switcher is expanded and for tooltips
     */
    fullLabel?: string;
    /**
     * The group the visualization belongs to
     */
    groupLabel: string;
    /**
     * The priority of the visualization in the list (global priority)
     * Higher number means higher priority. When omitted defaults to 0
     */
    sortPriority?: number;
    /**
     * Indicates if visualization is in the experimental stage.
     */
    showExperimentalBadge?: boolean;
}
export interface CellValueAction {
    id: string;
    type?: string;
    iconType: string;
    displayName: string;
    execute: (data: CellValueContext['data']) => void;
}
export type LayerCellValueActions = CellValueAction[][];
export type GetCompatibleCellValueActions = (data: CellValueContext['data']) => Promise<CellValueAction[]>;